HardwarePixelBuffer  - Fixed invalid XML comment PixelFormatFlags - Added missing [Flags] attribute PixelConverter - Fixed incorrect casting of ColorEx to (uint) in PackColor() Examples-DynTex.material - Removed SceneBlend on DynaTex3 so that the fish shows up

+++ Projects/Axiom/Source/Engine/Media/PixelConverter.cs
@@ -754,7 +754,7 @@ namespace Axiom.Media
 
 			unsafe
 			{
-				PixelConverter.PackColor( (uint)color.r, (uint)color.g, (uint)color.b, (uint)color.a, format, (byte*)( dest.ToPointer() ) );
+				PixelConverter.PackColor( color.r, color.g, color.b, color.a, format, (byte*)( dest.ToPointer() ) );
 			}
 		}
